Backend/pyproject.toml
[project]
name = "email-pdf-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chainlit>=2.5.5",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.58",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.1",
    "psycopg>=3.2.7",
    "psycopg-pool>=3.2.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
